Bug fixes
Lados das portas não deixam mais a mira vermelha
Substituidas algumas lampadas por apagadas lv1
fix posição do impacto laser da piranha
Aumetado vida do player em +2 e escudo em +1
Fix efeito de sangue na tela sumindo quando não devia
Fix Vida podendo ficar acima do maximo
Melhoria na mira da arma (atinge com maior precisão o centro da tela)
Tiros não colidem mais com items

+++ TCC Alien Shooter Project/Assets/Scripts/Canvas/CanvasManager.cs
@@ -78,10 +78,11 @@ public class CanvasManager : MonoBehaviour
         shieldImage.fillAmount = shield;
         healthImage.fillAmount = health;
         damageEffect.weight = (ElevateBy(1 - shieldImage.fillAmount, 2)) * 0.7f;
+
         var dmgPercent = ElevateBy(1 - health, 2);
         for (int i = 0; i < bloodSplaters.Length; i++)
         {
-            var indexPercent = ((float)i + 1) / bloodSplaters.Length;
+            var indexPercent = ((float)i + 1) / (float)bloodSplaters.Length;
             bloodSplaters[i].gameObject.SetActive(indexPercent <= dmgPercent);
         }
     }
